+++ src/constants/index.ts
@@ -12,3 +12,5 @@ export * from './vital-signs';
 export * from './network';
 export * from './sectors';
 export * from './smart';
+export * from './service-capabilities';
+export * from './verifiable-credentials';