+++ src/examples/contract-examples.ts
@@ -9,6 +9,7 @@
  */
 
 export * from './shared';
+export * from './ica-activation-proof';
 export * from './organization-controller';
 export * from './individual-controller';
 export * from './professional';